Break test Equatable model transform out to its own file
